tons of training updates
