+++ gwak/train/train.smk
@@ -6,6 +6,16 @@ cl_configs = [
     'S4_SimCLR_multiSignalAndBkg',
     'Transformer_patch_SimCLR_multiSignalAndBkg',
     'Transformer_patch_noClass_SimCLR_multiSignalAndBkg',
+    's4_kl0.5_bs512',
+    's4_kl1.0_bs512',
+    's4_kl2.0_bs512',
+    's4_kl0.5_bs256',
+    's4_kl1.0_bs256',
+    's4_kl2.0_bs256',
+    's4_kl1.0_bs256_noGlitch',
+    'transformer_patch64_kl0.5_bs512',
+    'transformer_patch64_kl1.0_bs512',
+    'transformer_patch64_kl2.0_bs512'
     ]
 fm_configs = [
     'NF_onlyBkg',
